+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## [Unreleased]
 
+- Automatically reset the `WaitingThreads` counter when enabling it (#7).
+- Disallow resetting the `WaitingThreads` counter when it is active (#7).
+
 ## [0.2.0] - 2023-03-28
 
 - Use C11 atomics instead of MRI's `rb_atomic_t`.